does not read 1D numpy arrays
The wav file that I was trying to convert returned a one dimensional numpy array and it was showing an error because a 1D numpy array does not show the number of rows as 1. 
Because of which on line 122: M, numChannels = data.shape  will show an error saying that it expected 2 arguments but it only got one. 


So this is a slight modification to the code. 
Thanks for the upload.